requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Flask==2.0.2
ibm-cloud-sdk-core==3.13.2
ibm-watson==5.3.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.5
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
websocket-client==1.1.0
Werkzeug==2.0.2
